# docs/swagger.yaml
basePath: /
info:
  contact: {}
  description: 轻量级网盘服务 API — 文件上传/下载、分片上传断点续传、用户认证、秒传去重、AI 语义检索
  title: gofile API
  version: "1.0"
paths:
  /file/download:
    get:
      description: 支持 Range 请求头实现断点续传和拖动播放
      parameters:
      - description: 文件 SHA1
        in: query
        name: filehash
        required: true
        type: string
      - description: Range 请求头（如 bytes=0-1023）
        in: header
        name: range
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 完整文件内容
        "206":
          description: 区间响应
        "404":
          description: 文件不存在
      security:
      - ApiKeyAuth: []
      summary: 下载文件
      tags:
      - 文件
  /file/query:
    get:
      description: 支持分页，无分页参数时返回全部文件
      parameters:
      - description: 页码（从 1 开始）
        in: query
        name: page
        type: integer
      - description: 每页数量（1-100，默认 20）
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 文件列表
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  code:
                    type: integer
                  data:
                    properties:
                      list:
                        type: array
                      page:
                        type: integer
                      size:
                        type: integer
                      total:
                        type: integer
                    type: object
                  msg:
                    type: string
                type: object
            type: object
      security:
      - ApiKeyAuth: []
      summary: 查询文件列表
      tags:
      - 文件
  /file/upload:
    post:
      consumes:
      - multipart/form-data
      description: 支持秒传去重（通过 filehash 参数），危险文件类型（.html/.js/.exe 等）会被拒绝
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: 文件 SHA1（秒传检测）
        in: formData
        name: filehash
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  code:
                    type: integer
                  data:
                    properties:
                      filehash:
                        type: string
                    type: object
                  msg:
                    type: string
                type: object
            type: object
        "400":
          description: 参数错误或文件类型不允许
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  code:
                    type: integer
                  data:
                    type: object
                  msg:
                    type: string
                type: object
            type: object
      security:
      - ApiKeyAuth: []
      summary: 上传文件
      tags:
      - 文件
  /user/info:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  code:
                    type: integer
                  data:
                    type: object
                  msg:
                    type: string
                type: object
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - 用户
  /user/logout:
    post:
      description: 删除服务端 token 并清除客户端 Cookie
      produces:
      - application/json
      responses:
        "200":
          description: 已退出登录
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  code:
                    type: integer
                  data:
                    type: object
                  msg:
                    type: string
                type: object
            type: object
      summary: 退出登录
      tags:
      - 用户
  /user/signin:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 验证用户名密码，成功后通过 Set-Cookie 设置 token（HttpOnly）
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  code:
                    type: integer
                  data:
                    properties:
                      Location:
                        type: string
                      Username:
                        type: string
                    type: object
                  msg:
                    type: string
                type: object
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  code:
                    type: integer
                  data:
                    type: object
                  msg:
                    type: string
                type: object
            type: object
      summary: 用户登录
      tags:
      - 用户
  /user/signup:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 注册新用户，密码需至少 8 位且包含大写/小写/数字/特殊字符中的至少三类
      parameters:
      - description: 用户名（至少 3 位）
        in: formData
        name: username
        required: true
        type: string
      - description: 密码（至少 8 位，含大写/小写/数字/特殊字符至少三类）
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  code:
                    type: integer
                  data:
                    type: object
                  msg:
                    type: string
                type: object
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              allOf:
              - {}
              - properties:
                  code:
                    type: integer
                  data:
                    type: object
                  msg:
                    type: string
                type: object
            type: object
      summary: 用户注册
      tags:
      - 用户
securityDefinitions:
  ApiKeyAuth:
    description: Cookie 鉴权（username + token），登录后由 Set-Cookie 自动设置
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"